(RE-550) Add debian and redhat packaging artifacts
This commit adds the basic packaging for pe-razor-server. The redhat packaging
is ready to roll, while the debian packaging will need updates to function once
razor has left the tech preview phase. The redhat packaging includes an init
script and a sysconfig file. The pe-razor-libs package will create the pe-razor
user which is used in the packaging to own the logs and config.ru.