(RE-550) Add and use jruby_vendor_gems pretar task
This commit defines a rake task called jruby_vendor_gems. The task installs the
gems listed in the Gemfile before the tarball is created so that the gems can
remain dynamic and not in the repository without adding the need to make remote
gem calls in the packaging build root.