Update bundle config and install the file
Previously an incorrect bundler setting was being applied to the bundler
config. This didn't matter, because the bundler config wasn't being included in
the package. This commit corrects both of these mistakes, by using the PATH
instead of --path bundler config setting, and by copying the .bundle directory
into the buildroot.