(PE-2345) Lay down sample configfile in /etc
Previously the sample configfile was laid down in
/opt/puppet/share/razor-server, which wouldn't be very intuitive for someone
looking to edit or create a configfile. This commit updates the rpm packaging
to instead put it in /etc/puppetlabs/razor, which is where pe-razor-server
expects to find a configfile. It also creates the /etc/puppetlabs/razor
directory, which wasn't previously created by the packaging.