Load RAZOR_CONFIG in the razor-admin wrapper for easier use
Without this, in order to successfully use the razor-admin wrapper a user would
need to export RAZOR_CONFIG to point to /etc/puppetlabs/razor/config.yaml.
RAZOR_CONFIG must be defined in /etc/sysconfig/pe-razor-server in order for the
service to start successfully, so this commit simple sources that file to get
RAZOR_CONFIG populated. After this, the razor-admin wrapper can be invoked
without exporting variables.