(RE-703) Deploy razor to torquebox via packaging
Previously razor-server was never "deployed" to torquebox. This commit adds
some pre, preun, and post scriptlets to handle deploying, undeploying and
migrating the database for razor.

On initial install, the application is deployed in the post so that torquebox
can run it. On upgrades, the application is first undeployed, then if the razor
config is present, the db is migrated, and then the application is deployed
again. On uninstall, the application is undeployed.