Maint: (Razor) Fix failing razor-client installation
The razor-client installation was resulting in an error: "undefined method
`map' for Gem::Specification:Class" which seems to be from out-of-date RubyGems.
This fix updates those before running the command. Additionally, there was a
check for the ok_json JSON adapter, but this is no longer present with this
Beaker run, so that check has been made conditional.