(RAZOR-387) Use appropriate manifest file for both installs and upgrades
When upgrading from an older PE version, we do not change the user's
manifest configuration. This means that during an upgrade the manifest
is 'site.pp' whereas in a fresh 3.7 install we use the manifest
directory setup. This change makes the razor installation step aware
of this difference so it can work in both cases.