(QA-1730) Create Prototype Integration Test for Razor
Added a new top-level folder for integration tests which are designed to test
Razor from end-to-end. The first integration test is for CentOS 6 provisioning
using a PE broker. This integration suite requied special intrastructure to be
created and configured. This prototype makes many assumptions and repeats
code in many places. This prototype is meant to be an experiment is verifying
that end-to-end tests are possible with Razor.