(re-4175) Update config variable definitions for el-7/systemd
Currently, /etc/sysconfig/pe-razor-server is not installed on
el-7/systemd systems, but is required by the razor-admin script.

This update adds the /etc/sysconfig/pe-razor-server file on el-7,
and updates it to be compatible with the format expected by systemd,
while maintaining backwards compatibility with SysVinit.

This also updates the systemd unit to start after pe-postgresql,
if pe-postgresql is also configured to start on the same system.