(maint) Export sysconfig variables from the razor-admin wrapper
We no longer explicitly export the variables set in the /etc/sysconfig/pe-razor-server
file, due to the way systemd handles those variables. This is causing
the razor-admin process, called by using the bash 'exec' from the wrapper,
to not pick up the variables set in that file.

This wraps the call that sources the sysconfig file with the
bash option to auto-export variables.