(maint) export sysconfig variables from the init script
We no longer explicitly export the variables set in the /etc/sysconfig/pe-razor-server
file, due to the way systemd handles those variables. This is causing
the razor-server process, when called by using the bash 'exec' from the init script,
to not pick up the variables set in that file.

This wraps the call that sources the sysconfig file with the
bash "-a" option to auto-export variables.