(maint) Fix acceptance test to use /opt/puppetlabs/puppet/bin/gem
CentOS ships with Ruby 1.8.7 and gem 1.3.7 installed by default. This fails
to run the client properly for a few reasons, but since these versions of
ruby are past their end-of-life, they will not be supported.

The fix is to install the razor-client using the version of gem that ships with
puppet.