More clear logging for hooks
Several cases around logging hook execution failure were improperly
handled. Severity, for instance, was reported both on the event log and
in the event log's json `entry`, potentially differing in value. Previously,
the STDERR output was ignored; this will be included as an error and affect
the execution's severity.

For https://tickets.puppetlabs.com/browse/RAZOR-642