Delete repo directory when repo is deleted
When the user runs `razor delete-repo --name some-repo`, the intent is to
delete the folder if the user supplied an `iso-url` property. That was, in many
cases, not happening because many ISO archives contain files which are
read-only, so the secure delete attempt fails.

The fix for this is to update the contents of the repo directory to allow write
privileges so they can be deleted properly.

This manifested when the user tried to create a new repo with the same name.
The archive extraction would fail from libarchive, and the error was "Can't
unlink already-existing object", likely from the same permissions reason.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-757