Handle relative directory when running hooks
Whenever a hook would run, the relative directory would be changed to the
hook's directory. That's normally not an issue, but in for example spec
tests, an accurate relative directory may be relied upon.

The fix is to simply restore the old working directory after the hook
execution completes.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-928