(PE-21066) Do not auto provide or require things
This commit prevents rpm from automatically adding certain dependencies
to this package. Without this change, we have a dependency on
`/usr/bin/ruby`, which pulls in system ruby. We really don't want this
to happen. By adding `AutoReq: 0`, we stop that from happening. This
commit also adds `AutoProv: 0` just for good measure, so that rpm
doesn't try to satisfy other dependencies with libraries we provide in
this package.